Add Bearer token parameter
for an external (Aria\VCF) Orchestrator to (Aria\VCF) Automation cannot authenticate to the Orchestrator. However if you create a Bearer token on the Automation that the Orchestrator is configured against you can use that Bearer token to Auth to Orchestrator.
Added the Bearer token as a third authentication method.